refactor(portal): drain Actions GraphQL docs to scenes/common/actions
