+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/Actions/ActionId/Danger/ActionDangerZoneContent/index.tsx
@@ -13,12 +13,12 @@ import { useUser } from "@auth0/nextjs-auth0/client";
 import { useRouter } from "next/navigation";
 import { useCallback, useMemo, useState } from "react";
 import { toast } from "react-toastify";
-import { GetActionsDocument } from "../../../page/graphql/client/actions.generated";
-import { GetSingleActionQuery } from "../page/graphql/client/get-single-action.generated";
-import { useDeleteActionMutation } from "./graphql/client/delete-action.generated";
+import { GetActionsDocument } from "@/scenes/common/actions/graphql/client/actions.generated";
+import { GetSingleActionForDangerQuery } from "@/scenes/common/actions/graphql/client/get-single-action-for-danger.generated";
+import { useDeleteActionMutation } from "@/scenes/common/actions/graphql/client/delete-action.generated";
 
 export const ActionDangerZoneContent = (props: {
-  action: GetSingleActionQuery["action_by_pk"];
+  action: GetSingleActionForDangerQuery["action_by_pk"];
   teamId?: string;
   appId?: string;
 }) => {
