+++ web/scenes/common/actions/graphql/client/app.generated.ts
