+++ web/scenes/common/actions/graphql/client/app.graphql
