+++ web/scenes/common/actions/graphql/client/delete-action.generated.ts
