+++ web/scenes/common/actions/graphql/client/fetch-action-stats.gql
