+++ web/scenes/common/actions/graphql/client/get-action-name.generated.ts
