+++ web/scenes/common/actions/graphql/client/get-action-name.graphql
