+++ web/scenes/common/actions/graphql/client/get-kiosk-action.generated.ts
