+++ web/scenes/common/actions/graphql/client/get-single-action.generated.ts
