+++ web/scenes/common/actions/graphql/client/toggle-kiosk.graphql
