+++ web/scenes/common/actions/graphql/server/get-action-insert-permissions.graphql
