+++ web/scenes/common/actions/graphql/server/get-action-update-permissions.graphql
