+++ web/scenes/common/actions/graphql/server/insert-action.generated.ts
